Migrated away from using external anime data.

+++ api/internal/models/dtos/userAnimeDetails.go
@@ -1,7 +1,7 @@
 package dtos
 
 type UserAnimeDetailsDto struct {
-	Rating             int64  `json:"rating"`
+	Rating             int  `json:"rating"`
 	WatchStatus        string `json:"watchStatus"`
-	NumEpisodesWatched int64  `json:"numEpisodesWatched"`
+	NumEpisodesWatched int  `json:"numEpisodesWatched"`
 }
